Move session runtime files behind a module boundary
This split keeps the behavior from main unchanged while giving session-owned concepts a folder boundary before history and persistence code are added in the stacked PR.\n\nConstraint: No history behavior changes in this base PR\nRejected: Add history while moving files | makes the stacked review harder to read\nConfidence: high\nScope-risk: narrow\nTested: pnpm run test; pnpm run typecheck
