+++ src/index.ts
@@ -1,6 +1,5 @@
 export { Agent } from "./runtime/agent";
 export { runAgentLoop } from "./runtime/agent-loop";
-export type { AgentEvent, AgentEventListener } from "./runtime/events";
 export { createMockLlm, mockLlm } from "./runtime/mock-llm";
 export type { Llm, LlmContext, LlmOutput, LlmOutputPart } from "./runtime/mock-llm";
-export type { AgentSession, SessionInput } from "./runtime/session";
+export type { AgentEvent, AgentEventListener, AgentSession, SessionInput } from "./runtime/session";
